Added support for tower angle trim values
This change is based on the work from Rich Cattell to add support for
adjusting tower angles in Marlin.
    https://github.com/RichCattell/Marlin

It appears that his branch is no longer being maintained, and current
Marlin has implemented some portions such as the rod radius adjustments
but not the angle adjustments.

This change attempts to reproduce the angle trim support only. No other
changes have been implemented. I have also attempted to do it in a way
that matches the current Marlin firmware as closely as possible in hopes
that it will prove easier to maintain.